\name{rppaDesigner}
\alias{rppaDesigner}
\alias{rd}
\title{Graphical User Interface for the SuperCurve package}
\description{
  Graphical User Interface for the SuperCurve package.
}
\usage{
rd()
}
\details{
  The function \code{rd} launches a Graphical User Interface
  for recording the layout of positive and negative controls in an RPPA
  experiment.  The GUI uses Tk widgets (via the R TclTk
  interface by Peter Dalgaard) in order to provide a simple interface.

}
\author{P. Roebuck <plroebuck@mdanderson.org>, Kevin R. Coombes}
\examples{
  \dontrun{
rd()
  }
}
\keyword{IO}

