% $Id$

\name{SuperCurveGUI-package}
\alias{SuperCurveGUI-package}
\docType{package}
\title{GUI for reverse phase protein lysate array analysis}
\description{
  A package for analyzing reverse phase protein lysate arrays (RPPA).
}
\details{
  \tabular{ll}{
    Package: \tab SuperCurveGUI\cr
    Type:    \tab Package\cr
    Version: \tab \Sexpr[results=text]{packageDescription("SuperCurveGUI")$Version}\cr
    Date:    \tab \Sexpr[results=text]{packageDescription("SuperCurveGUI")$Date}\cr
    License: \tab \Sexpr[results=text]{packageDescription("SuperCurveGUI")$License}\cr
  }

  For a complete list of functions, use \code{library(help="SuperCurveGUI")}.\cr
  For a high-level summary of the changes for each revision, use
    \code{file.show(system.file("NEWS", package="SuperCurveGUI"))}.
}
\author{
  P. Roebuck \email{proebuck@mdanderson.org}
}
\keyword{package}

