\name{SuperCurveGUI-package}
\alias{SuperCurveGUI-package}
\docType{package}
\title{GUI for reverse phase protein lysate array analysis}
\description{
  A package for analyzing reverse phase protein lysate arrays (RPPA).
}
\details{
  \tabular{ll}{
    Package: \tab SuperCurveGUI\cr
    Type: \tab Package\cr
    Version: \tab 1.0-0\cr
    Date: \tab 2008-06-06\cr
    License: \tab See LICENSE for license information\cr
  }

  For a complete list of functions, use \code{library(help="SuperCurveGUI")}.
  For a high-level summary of the changes for each revision, use
    \code{file.show(system.file("NEWS", package="SuperCurveGUI"))}.
}
\author{P. Roebuck, \email{plroebuck@mdanderson.org}}
\keyword{package}

