\name{supercurveGUI}
\alias{supercurveGUI}
\title{Graphical User Interface for the SuperCurve package}
\description{
  Graphical User Interface for the SuperCurve package.
}
\usage{
supercurveGUI()
}
\details{
  This function launches a Graphical User Interface for the SuperCurve package.   The GUI uses Tk widgets (via the R TclTk interface by Peter Dalgaard) in
  order to provide a simple interface to the SuperCurve functions for analyzing
  reverse phase protein lysate arrays.
}
\author{P. Roebuck <plroebuck@mdanderson.org>}
\examples{
  \dontrun{
supercurveGUI()
  }
}
\keyword{IO}

