\name{supercurveGUI}
\alias{supercurveGUI}
\alias{sc}
\title{Graphical User Interface for the SuperCurve package}
\description{
  Graphical User Interface for the SuperCurve package.
}
\usage{
supercurveGUI()
sc()
}
\details{
  The function \code{supercurveGUI} launches a Graphical User Interface
  for the SuperCurve package. The GUI uses Tk widgets (via the R TclTk
  interface by Peter Dalgaard) in order to provide a simple interface
  to the SuperCurve functions for analyzing reverse phase protein lysate
  arrays. The function \code{sc} is an alias for the aforementioned function,
  provided for backwards compatibility.
}
\author{P. Roebuck \email{plroebuck@mdanderson.org}}
\examples{
  \dontrun{
supercurveGUI()
  }
}
\keyword{IO}

